make the busy timer public
the busy timer is used to measure the time spent working on a request
outside of waiting for external calls like coprocessors and subgraph
calls. It still includes the time spent waiting for other concurrent
requests to be handled (wait time in the executor) to show the actual
router overhead in the request handling.
This makes the busy timer API public to let native plugins use it when
they do their own network calls.