fix rust version in CI (#4997)
When building an old version of the Router, we have a
`rust-toolchain.toml` file that fixes the version used to build, but
`rustup.sh` does not actually take that file into account: it installs
the last stable, that is then cached, and then when building the router,
we install on the fly the required version. This is not a big issue when
working on the `dev` branch which always uses the last stable. But for
older versions, we should cache the version actually used