add conditions on custom attributes for spans + a new selector for graphql error (#4987)
Currently the router does not set otel.status_description on spans. This
is a special attribute that will also set the span status.

This means that even responses have errors or http status is not 200
spans are not flagged as an error.

To complicate matters further, other languages usually use exceptions
and stack unwinding to set span status, both of which are not available
to us in rust.



Example:

```yaml
telemetry:
  instrumentation:
    spans: 
      router: 
        attributes:    
          otel.status_description: 
            static: "there was an error"
            condition:
              any:
              - not:
                  eq:
                  - response_status: code
                  - 200
              - eq:
                - on_graphql_error
                - true
```

This will allow the user to set the description to something if there
was an issue.

Suggest implementing using the existing condition logic and add two new
selectors. This would check http status and if there are 1 or more
graphql errors.

Fixes #4336

---------

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>
Co-authored-by: bryn <bryn@apollographql.com>