Use the entire schema when hashing an introspection query (#5007)
Fix #5006

in https://github.com/apollographql/router/pull/4883 we introduced a
query hashing scheme that stays stable across schema updates if the
update does not affect the query. Unfortunately, it was not taking
introspection queries into account.
This fixes the hashing mechanism to add the schema string to hashed data
if we encounter an introspection field. This is a temporary fix until we
move introspection execution out of query planning. At that point, this
hashing mechanism won't ever see introspection fields.