tests: Remove "is proto fresh" test in lieu of GitHub Action
We'll have a GitHub Action do the maintenance on this file, rather than
having a test remind us that it's stale.  The test results in all PRs (and
all downstream users of this repository) suddenly getting failing tests
every time our infra-team deploys a new version of the proto, whereas the
GitHub Action can be a much more controlled burn.

The GitHub action that replaces this is in
https://github.com/apollographql/router/pull/5030, and we will merge this
test just as soon as that is validated.