chore/nit: Fix warnings and trailing newlines in CircleCI Config
In practice, this file actually serves as a cache key for our configuration
due to the incorporation of the file's hash in the [key/ID], which I need to
bump in order to rotate the cache in CircleCI.  While I'm here, I might as
well fix the editor warnings being emitted about this not being a valid
version of Xcode, along with some trailing newlines!  While we had '14.2' as
the value to bring in patch releases of Xcode, I'd rather we not have that
level of indeterminism anyhow, so I'll just make it '14.2.0' which is the
latest '14.2' release anyhow.

[key/ID]: https://github.com/apollographql/router/blob/0137e0cf848b9b22ef94fe14b32ed1a082d3f0b5/.circleci/config.yml#L73-L83