generate the JSON schema only once
Surprisingly, the most expensive part of running tests in the router is
in JSON schema compilation. Every time we validate a configuration, we
generate the JSON configuration schema (which is about 31k lines at this
point) then compile a JSONSchema structure used to validate the
configuration.
This computation can be done only once per execution without losing
anything, as it only depends on the configuration schema, which is
generated from the router's code and native plugins.