Remove the potential to panic during operation parse/validate
The possibility that an ExecutableDocument::parse_and_validate() will
fail, means there is a possibility that a router will panic when
`as_parsed()` or `as_serialized()` is invoked.

We remove the potential to panic by logging errors for any detected
errors and continuing to use the Valid (but incomplete) document, which
is the behaviour we had before this change.

Summary: This restores the preceding behaviour of the router with the
additional error logging of detected errors.