chore(tests): provide more threads to each integration test in nextest config (#5109)
Nextest by default will try to run a test on each CPU core available. However this isn't suitable for integration tests as the router will be using multiple threads. This is leading to many test failures in CI.

In addition the scaffold test really shouldn't be run at the same time as any other test as it invokes rustc. This manifests as tests taking 3 minutes or more if they are unlucky enough to get scheduled at the same time.