Fix reduce()
Consider a fetch dependency graph with nodes A, B, C and edges 1, 2, 3:

```
A -1> B -2> C
 \----3--->/
```

`reduce()` should remove edge `3` as you can already reach `C` by
following edge `1` and `2`.

The bug in this code was instead removing edge `2` in this situation. We
have to go through grandchildren and all nested children of the
grandchildren for each node and remove edges to the original node.