Fix semantic merge conflict with https://github.com/apollographql/router/pull/5088
PR #5088 changed the signature of a method.
It had a green CI run and git succedded at squashing/rebasing it,
but the target `dev` branch gained a new call to that method in the mean time
so the resulting merge has a compile error.

This would have been prevented if we ran CI on exactly what was about to
be pushed to `dev`, instead of only on PR branch.
The Rust project calls this [the "not rocket science" rule
of software engineering](https://graydon.livejournal.com/186550.html).