Adjust privacy of JSONSelection struct internals.
Resolves Jira issue CNN-152.
Closes #5249.

Although keeping the JSONSelection structures immutable would have been
(and still may be) ideal, we currently have a few algorithms that rely
on incrementally constructing those structures (specifically populating
SubSelection with new NamedSelection items), so we need to support a
limited form of mutability.

To make that mutability safe(r), I decided to replace all direct access
to struct fields with public methods, which enforce the invariant that
NamedSelection items may only be _appended_ to a &mut SubSelection. This
enforcement works by making the fields of the SubSelection struct (and
other JSONSelection-related structs) pub(super), which means they are
only accessible and constructible within the json_selection directory
(which is important so the ApplyTo trait can still access them).

Since SubSelection is now guaranteed to be append-only, I was also able
to expose a SubSelection::index_of_named_selection method, which returns
an unforgeable and permanently valid NamedSelectionIndex wrapper struct,
allowing safe and infallible access via SubSelection::get_at_index and
SubSelection::get_at_index_mut. This indirection is helpful for
placating the borrow checker, which refuses to accept any variation of a
method like

   SubSelection::get_mut(&mut self, i: usize) -> Option<&mut NamedSelection>

as a replacement for subselection.selections.get_mut(i), due to the
complexities of working with exclusive mutable references.

It might have been nice to restrict the constructability of enum
variants as well (especially for PathSelection), but there's no way to
do that in Rust without also removing the ability to match against the
variants, which seems too onerous (and difficult to replicate with
public functions).