chore(federation): refactor RebasedFragments
`RebasedFragments` is a cache of fragments rebased on different
subgraphs. It's only used internally in the query planner and isn't
really generally useful, so this moves the structure out of the general
`operation` module and into the `query_plan` module.

`RebasedFragments` would be cloned to do recursive planning for
conditions. This would mean multiple Arc references to the rebased
cache would be alive, and calls to `for_subgraph` would clone the cache.
Different instances of `RebasedFragments` are independent, so the cache
could be out of sync and cause duplicate work.

This changes `RebasedFragments` to be passed by reference, so it's never
cloned. It uses runtime borrow-checking to fill the cache.
`RebasedFragments` instances are only used for one `build_query_plan`
call, which is fully synchronous and single-threaded, so we can
guarantee that conflicting borrows do not happen.