chore(docs): Replicate custom histograms CHANGELOG to v1.46.0.
In theory v1.44.0 was the first version in which this was fixed, but we
don't recommend running that version for the reasons that are already noted
in the CHANGELOG.md.

Therefore, the first version that we recommend using to utilize custom
historgrams is v1.46.0.  This updates the CHANGELOG.md to signal that with a
note and a duplication of the original CHANGELOG entry.

Original Issue: https://github.com/apollographql/router/issues/4543
Initial PR: https://github.com/apollographql/router/pull/4572
Stabilizing PR: https://github.com/apollographql/router/pull/4771