Move QP "both" mode comparison to a background thread with bounded queue (#5341)
Before this PR on query plan cache miss we would, sequentially:

* Run the Rust QP
* Run the JS QP
* Compare results
* Return the result from JS

After this PR, in the main code path:

* Run the JS QP
* Spawn a background job with that result (shared in Arc) and the input document
* Return the result from JS

Meanwhile, the Rust QP and the result comparison runs in the background on a dedicated thread pool. Jobs are dropped if the bounded queue is full. To limit resource consumption at the moment the "pool" is a single thread, and the queue size is 10. Both of these parameters are const items in the code (not exposed in Router YAML config at this time).