When processing selectors for graphql instruments we need to make sure we perform no heap allocation otherwise performance is affected.
This PR modifies the `on_response_field` to take a Vec instead of returning a Vec allowing the same vec to be reused over and over.

When processing a field attributes from selectors are appended to this vec, and once the instrument has been updated the vec is truncated to its original state.

In addition, we don't need to look at attributes unless the instrument is actually going to get updated, so we also shift the attribute processing to after we have checked that there is an update to be done.