Implement evaluated_plan_count statistic
Fixes most of FED-272

To avoid the risk of incrementing a cloned counter and then dropping it,
this PR removes `derive(Clone)` on `QueryPlanningStatistics` and
pass it around by reference instead.

I tried make this a `&mut QueryPlanningStatistics` exclusive reference,
but that would require changing every use of `&QueryPlanningParameters`
to be a `&mut` exclusive reference too.
This proved difficult as parameters are shared in multiple places.

Instead, the counter is now a `Cell<usize>` so that it can be incremented
through a shared reference.
This causes `QueryPlanningStatistics` and `QueryPlanningParameters`
to be `!Sync` but that’s ok: they only exist within one single-threaded
call to `build_query_plan()`.

The `QueryPlanner` itself is still `Sync` and multiple threads can run
separate `build_query_plan()` calls.

*