chore(federation): remove `FederationSchema::schema_mut`
`FederationSchema` requires that the schema isn't modified behind its
back. Otherwise the referencers may get out of sync.

Only API schema was using this function. It doesn't use it in a way that
would affect referencers, but the method should still not exist. Here
I replaced it by unwrapping the schema and then re-wrapping it after API
schema has done its thing. It means referencers will be re-computed, but
all in all that's not the biggest issue in the world.