add support for query metrics to error response and update snapshots
The changes to add support for query metrics appear to be wide ranging
in terms of snapshot impact. Some of the snapshots are so large I'm not
sure if I've done the right thing in accepting the change...

I've also added support for query metrics to the error path, but I'm not
sure if or how that should be documented.