Re-work storage of query metrics
Rather than holding the data in a Context on the router bridge, keep the
measured information directly on the QueryPlan. This has the advantage
that the data will be available for queries that are cached.

This fixes a variety of unusual failures in the varous batch tests,
since caching was causing some very odd failure cases.