fix(federation): pass fragments when normalizing selections (#5527)
When normalizing the selection sets they might contain references to named fragments. We need to pass valid named fragments in order to normalize.

Note that in the JS codebase, fragments kept reference to named fragments hence when rebasing it would work without explicitly passing the fragments. In RS we don't store those references so when we rebase/normalize we have to explicitly pass them in.