chore(federation): implement Deref on keyed operation types
These are types that encode a selection key together with immutable data
for that selection. The purpose of these types is that you can't mutate
the things that are encoded in the key.

To do this the `Field`, `FragmentSpread`, and `InlineFragment` types hid
the data behind a `.data()` accessor. But all it does is ensure that the
data is immutable: we can do the same with deref, without having to call
`.data()` everywhere.

The `.data()` accessor still exists because in some places we
specifically clone only the data and not the key.