chore(tests): Apply nextest retries rules to known-flaky tests.
This attempts to apply retries to known flaky tests to provide some
air-cover in CI while we action the specific reasons.  This choses to only
retry known failures to avoid introducing net-new flaky tests inadvertently.

To get those known flaky teses, we used the CircleCI API flaky_tests API to
generate it.  I called the CircleCI using this below curl command,
basically, but you'll need to set the 'VALUE_FROM_YOUR_COOKIES' for the
'ring-session' correctly, and you'll need to get that from your CircleCI
browser cookies after logging in:

    curl 'https://circleci.com/api/v2/insights/github/apollographql/router/flaky-tests' \
      -H 'accept: */*' \
      -H 'accept-language: en-US,en;q=0.9' \
      -H 'cookie: ring-session=VALUE_FROM_YOUR_COOKIES' |
        jq -r '.flaky_tests[] | "or ( binary_id(=\(.classname)) & test(=\(.test_name)) )"' | sort | uniq

Then, put the output into the '.config/nextest.toml', but be sure to remove
the first 'or' from the output.