fix(federation): optimize `runtime_types_intersect`
This is an alternative to #5534 that does not add a new cache.

It does the intersection check with just the information that's readily
available on the schema.

`possible_runtime_types` is still slow, but it's no longer called in
this very hot path. If it turns up in profiling later we can still
consider optimizing `possible_runtime_types` with a cache or with
wrapper types.