Entity cache: set up basic infrastructure for invalidation (#5544)
This sets up the Notify structure to act as pub/sub for invalidation events. A task is started by the entity cache plugin to execute those events. There is apparently no way to mock Redis SCAN commands with the fred crate, so instead this modifies the integration tests to be able to send back more complex subgraph responses, with headers and extensions. It also adds matching on method, headers, path at the subgraph mock level.