chore(federation): remove excessive selection set re-building
In JS, this called `toSelectionSet()` on a `SelectionSetUpdates`. In
theory, doing `make_selection_set` is the Rust equivalent.

However, we are not working with a `SelectionSetUpdates`. In fact,
`optimized.add_selection` and `optimized.add_selection_set` already
handle merging duplicate selections correctly. So the call to
`make_selection_set` should not make any changes.