fix(federation): remove temporary string allocations from position `try_from()` errors
Much like `try_get()` in #5605, we regularly use `try_from()` on
positions just to check if we *can* convert, or to conditionally do
something with a converted type. In some hot paths like rebasing or
optimize(), we do not actually use the error. So creating a full error
with an allocated string and a backtrace is excessive in those cases and
ends up taking a lot of time.

This adds a cheap internal error type for `Position::try_from()`
conversion failures. I also added macros to declare the conversions
because they're all the same.