chore(federation): simplify `subselection_type_if_abstract`
This was using its own `AbstractType` type that is identical to but
worse than `AbstractTypeDefinitionPosition` (probably a historical
accident :).)

It also needlessly converted to an `element`, which is probably what the
JS did, but because the Rust code is structured a bit differently, this
required extra cloning only to get to information that was already in
the original structure. Now that no longer happens.