feat(xtask): add a shorthand to profile the planner
We are working through some known-bad operations, and this is a really
useful way to generate a flame graph quickly and investigate specific
queries.

```
cargo xtask fed-flame plan /path/to/query.graphql /path/to/supergraph.graphql
```

This compiles the apollo-federation CLI in release mode with debug info
enabled, and then runs it with [samply], to produce a call tree and a
flame graph.

[samply]: https://github.com/mstange/samply