Refactor PathSelection enum into struct containing private PathList.
Although it may seem like a minor thing, we need an outer PathSelection
wrapper around an inner PathList so we can apply certain logic only at
the beginning of the path, which was not previously possible using only
the PathSelection enum.

For example, during the recursion of pretty_print_with_indentation, we
don't know if a given PathList::Key is the first key in the path, in
which case it _might_ need a leading '.' (depending on is_single_key),
or some subsequent key, in which case it always needs a leading '.'. If
we make a distinction between PathSelection and PathList, we can easily
apply that special logic only at the beginning of the path in
PathSelection::pretty_print_with_indentation.

As a side benefit, we can keep the PathSelection::path field private,
since PathSelection is now a struct rather than an enum, which should
help maintain the various invariants we want to enforce upon the path
(e.g. variables can only come at the beginning, and selections can only
come at the end), because code outside the json_selection directory can
no longer create any arbitrary chain of PathSelection enum variants.