Datadog sampling priority not set
In #5609 The backported code from the datadog exporter used code from the feature `agent-sampling`.

This was an incorrect decision as this requires that the sampler is updated. to support such functionality resulting in traces not being shown in the Datadog UI even though they were exported.

This PR switches the code with `not(agent-samping)` code which will delegate sampling to the sampler in the router yaml config.

Relates to #5609