fix: include __typename in response selection
router response formatting will insert __typename in the appropriate places for client responses, but we also need __typename in REST API response mapping to create complete entity references.

this is an interim approach for cheaply resolving typenames in a selection mapping by hardcoding a special `$typenames` variable prefix. it works only for concrete types (which is ok because we don't support abstract types yet.)