chore(federation): Coerce hardcoded GraphQL values to match JS
This PR does a small amount of GraphQL value coercion to better support
mismatch testing with the JS query planner.

Now if you have a schema like this:
```graphql
type Query {
  getUsers(ids: [Int]): [User]
}
```
and you send a query that uses a single literal for the list argument:
```graphql
{ getUsers(ids: 1) }
```
The query planner outputs a subgraph query that does:
```graphql
{ getUsers(ids: [ 1 ]) }
```

It's valid to provide a single value where a list is expected. QP will
just translate it to an actual list, similarly to what graphql-js does.

We already have a graphql-js compatibility layer like this for the API
schema implementation. This PR reuses some of that code.