add metrics for rust qp init failures
This adds a function to capture metrics for rust qp intialisation. We want to capture whether or not initialisation was a success. We also want to capture which unsupported feature triggered a failure - `@context`, progressive `@overrides` or a fed1 supergraph. There is an addition of "internal init error" to capture the remaining possible match arms.

In order to classify unsupported features, this commit also adds an `UnsupportedFeatureKind` enum to apollo-federation errors.