Allow Alias to use Key (not just Identifier).
While we might like to promote the use of identifier output keys when
generating GraphQL-safe JSON, the JSONSelection language is useful for
purposes other than generating GraphQL responses, e.g. generating HTTP
POST request bodies, where it's entirely possible an output key might
need to be quoted because it's not an identifier.