Add a histogram metric tracking evaluated query plans
the `experimental_plans_limit` option can be used to limit the number of query plans evaluated for a query, but there was no way to know what is a reasonable value for that option. This adds a histogram metric tracking the number of evaluated query plans, which should give more context to configure this.