fix(federation): treat invalid subgraph operations as internal errors
Currently, if we produce an invalid subgraph operation, it looks exactly
the same as a validation error from the input query. This makes issues
harder to debug.

This PR turns validation errors at the end of planning into internal
errors, so we know where they come from, and so end users do not get
confused about what a validation error means. The new error explicitly
mentions that it's a bug in the query planner that should be reported to
Apollo.

It's not possible to add a test for this, as it only triggers for bugs.
I manually tested it with the bug that's fixed in #5785:

```
Details: Query planning produced an invalid subgraph operation.
Error: fragSpreadOnly directive is not supported for INLINE_FRAGMENT location
```