chore(tests): Update flaky tests retry list from CircleCI API
Repeats the https://github.com/apollographql/router/pull/5577 motion

Overall, the API only shows tests that are repeatedly failing BEFORE
retriese are applied, so this is careful to not change the existing sorting
in generating the diff so we can see the new tests in the context of this PR
but still maintain the previously flaky list.  This was done with `sort | uniq`
with the existing list to interpolate the new tests.