Implement ApplyToInternal for Parsed<T> structures (not T).
In order to include location information in ApplyToError objects, the
apply_to_path method needs access to the Parsed<T> wrapper, not just the
wrapped T structure.

Thanks to the Parsed<T> struct implementing Deref, this change requires
remarkably few changes to the bodies of the apply_to_path methods.

A noteworthy exception to this new approach is PathSelection, which
always takes its location from self.path, so wrapping it with
Parsed<PathSelection> would be redundant.