fix(fed): corrected insertion logic in `add_at_path` (#5963)
In a recent change that added logic in `add_at_path` to remove unnecessary directives from selections, the insertion logic into the selection set had a bug. If the set did not have the key of the `OpPathElement`, we would generate a selection from the element after removing unnecessary directives. However, this was using the `entry` API, which requires that the key of the `Selection` passed to `.or_insert` matches the original. Because we removed the unnecessary directives, these keys were prone to mismatching.

Co-authored-by: Iryna Shestak <shestak.irina@gmail.com>