Rust intropsection: disable max depth rule
This protection against introspection queries generating huge responses
was added recently in graphql-js https://github.com/graphql/graphql-js/pull/4118
and ported to rust https://github.com/apollographql/apollo-rs/pull/904,
but is not yet present in the graphql-js version used by router-bridge.

This disables it for now from Rust introspection, in order to match
the current state of JS introspection.

Adding this rule (in both implementations) can be revisited separately.
In particular: the depth limit is hard-coded to 3. Is that the right number?
Should it be configurable? Is the rule checking the right set of fields?